App startup topic update
Update
