--- aspnetcore/fundamentals/startup/sample/.bowerrc
@@ -1,3 +1,0 @@
-{
-  "directory": "wwwroot/lib"
-}
