--- aspnetcore/fundamentals/startup/sample/Models/AppOptions.cs
@@ -1,7 +1,0 @@
-namespace StartupFilterSample.Models
-{
-    public class AppOptions
-    {
-        public string Option { get; set; } = "Option Default Value";
-    }
-}
