--- aspnetcore/fundamentals/startup/sample/Pages/Error.cshtml.cs
@@ -1,17 +1,0 @@
-using System.Diagnostics;
-using Microsoft.AspNetCore.Mvc.RazorPages;
-
-namespace StartupFilterSample.Pages
-{
-    public class ErrorModel : PageModel
-    {
-        public string RequestId { get; private set; }
-        
-        public bool ShowRequestId => !string.IsNullOrEmpty(RequestId);
-
-        public void OnGet()
-        {
-            RequestId = Activity.Current?.Id ?? HttpContext.TraceIdentifier;
-        }
-    }
-}
