--- aspnetcore/fundamentals/startup/sample/Pages/Index.cshtml
@@ -1,25 +1,0 @@
-﻿@page
-@model IndexModel
-@using Microsoft.Extensions.Options
-@inject IOptions<AppOptions> OptionsAccessor
-@{
-    ViewData["Title"] = "Startup Filter Sample";
-}
-
-<h1>@ViewData["Title"]</h1>
-
-<div class="row">
-    <div class="col-md-6">
-        <div class="panel panel-default">
-            <div class="panel-heading">
-                <h3 class="panel-title">AppOptions</h3>
-            </div>
-            <div class="panel-body">
-                <p>
-                    <b>@@OptionsAccessor.Value.Option:</b> 
-                    @OptionsAccessor.Value.Option
-                </p>
-            </div>
-        </div>
-    </div>
-</div>
