--- aspnetcore/fundamentals/startup/sample/Pages/Index.cshtml.cs
@@ -1,11 +1,0 @@
-﻿using Microsoft.AspNetCore.Mvc.RazorPages;
-
-namespace StartupFilterSample.Pages
-{
-    public class IndexModel : PageModel
-    {
-        public void OnGet()
-        {
-        }
-    }
-}
