--- aspnetcore/fundamentals/startup/sample/Pages/_ViewImports.cshtml
@@ -1,4 +1,0 @@
-@using StartupFilterSample
-@using StartupFilterSample.Models
-@namespace StartupFilterSample.Pages
-@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
