--- aspnetcore/fundamentals/startup/sample/appsettings.Development.json
@@ -1,10 +1,0 @@
-﻿{
-  "Logging": {
-    "IncludeScopes": false,
-    "LogLevel": {
-      "Default": "Debug",
-      "System": "Information",
-      "Microsoft": "Information"
-    }
-  }
-}
