--- aspnetcore/fundamentals/startup/sample/appsettings.Production.json
@@ -1,10 +1,0 @@
-﻿{
-  "Logging": {
-    "IncludeScopes": false,
-    "LogLevel": {
-      "Default": "Error",
-      "System": "Information",
-      "Microsoft": "Information"
-    }
-  }
-}
