--- aspnetcore/fundamentals/startup/sample/appsettings.json
@@ -1,8 +1,0 @@
-﻿{
-  "Logging": {
-    "IncludeScopes": false,
-    "LogLevel": {
-      "Default": "Debug"
-    }
-  }
-}
