--- aspnetcore/fundamentals/startup/sample/wwwroot/css/site.css
@@ -1,21 +1,0 @@
-body {
-    padding-top: 50px;
-    padding-bottom: 20px;
-}
-
-h1 {
-    font-size: 30px;
-}
-
-h2 {
-    font-size: 24px;
-}
-
-.body-content {
-    padding-left: 15px;
-    padding-right: 15px;
-}
-
-.panel-body {
-    font-size: 16px;
-}
