+++ aspnetcore/fundamentals/startup/sample_snapshot/Program2.cs
