+++ aspnetcore/fundamentals/startup/sample_snapshot/RequestSetOptionsMiddleware.cs
