+++ aspnetcore/fundamentals/startup/sample_snapshot/RequestSetOptionsStartupFilter.cs
