+++ aspnetcore/fundamentals/startup/sample_snapshot/Startup2.cs
